Have the user confirm first before deleting the saved data.

+++ js/main.js
@@ -614,6 +614,9 @@ function saveData() {
 }
 
 function deleteData() {
+    if (!confirm("This is gonna clear all your seats, are you sure?")) {
+        return
+    }
     localStorage.clear()
     location.reload()
 }
